Use size_t instead of int

+++ cppread/Data_block.cxx
@@ -118,8 +118,8 @@ namespace
 
 template <typename T>
 std::vector<T> Data_block::get_vector(const std::string& name,
-                                      const int length,
-                                      const int start_index)
+                                      const size_t length,
+                                      const size_t start_index)
 {
     std::vector<T> v(start_index + length);
     std::transform(data_series.at(name).begin(),
@@ -133,9 +133,9 @@ std::vector<T> Data_block::get_vector(const std::string& name,
 template <typename T>
 void Data_block::get_vector_range(std::vector<T>& destination,
                                   const std::string& name,
-                                  const int length,
-                                  const int source_start_index,
-                                  const int destination_start_index)
+                                  const size_t length,
+                                  const size_t source_start_index,
+                                  const size_t destination_start_index)
 {
     std::transform(data_series.at(name).begin()+source_start_index,
             data_series.at(name).begin()+source_start_index+length,
@@ -143,11 +143,11 @@ void Data_block::get_vector_range(std::vector<T>& destination,
             [](std::string value) { return convert_from_string<T>(value); });
 }
 
-template std::vector<std::string> Data_block::get_vector(const std::string&, const int, const int);
-template std::vector<double> Data_block::get_vector(const std::string&, const int, const int);
-template std::vector<int> Data_block::get_vector(const std::string&, const int, const int);
+template std::vector<std::string> Data_block::get_vector(const std::string&, const size_t, const size_t);
+template std::vector<double> Data_block::get_vector(const std::string&, const size_t, const size_t);
+template std::vector<int> Data_block::get_vector(const std::string&, const size_t, const size_t);
 
-template void Data_block::get_vector_range(std::vector<std::string>&, const std::string&, const int, const int, const int);
-template void Data_block::get_vector_range(std::vector<double>&, const std::string&, const int, const int, const int);
-template void Data_block::get_vector_range(std::vector<int>&, const std::string&, const int, const int, const int);
+template void Data_block::get_vector_range(std::vector<std::string>&, const std::string&, const size_t, const size_t, const size_t);
+template void Data_block::get_vector_range(std::vector<double>&, const std::string&, const size_t, const size_t, const size_t);
+template void Data_block::get_vector_range(std::vector<int>&, const std::string&, const size_t, const size_t, const size_t);
